@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:simple_dialogs_flutter/simple_dialogs_flutter.dart';
-
 void main() {
-
+  test('placeholder test', () {
+    expect(true, true);
+  });
 }
